move leagues fetch to server
